refactor: Added materials folder
