+++ src/materials/material.cpp
