+++ src/materials/material.h
@@ -1,10 +1,10 @@
 #ifndef MATERIAL_H
 #define MATERIAL_H
 
-#include "rt3.h"
-#include "surfel.h"
-#include "ray.h"
-#include "paramset.h"
+#include "../core/rt3.h"
+#include "../core/surfel.h"
+#include "../core/ray.h"
+#include "../core/paramset.h"
 
 namespace rt3 {
